[Silver III] Title: 주유소, Time: 320 ms, Memory: 35908 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -8,18 +8,18 @@ public class Main {
         BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(System.in));
 
         int N = Integer.parseInt(bufferedReader.readLine());
-        int[] len = new int[N - 1];
-        int[] price = new int[N - 1];
+        long[] len = new long[N - 1];
+        long[] price = new long[N - 1];
 
         StringTokenizer stringTokenizer1 = new StringTokenizer(bufferedReader.readLine());
         StringTokenizer stringTokenizer2 = new StringTokenizer(bufferedReader.readLine());
         for (int i = 0; i < N - 1; i++) {
-            len[i] = Integer.parseInt(stringTokenizer1.nextToken());
-            price[i] = Integer.parseInt(stringTokenizer2.nextToken());
+            len[i] = Long.parseLong(stringTokenizer1.nextToken());
+            price[i] = Long.parseLong(stringTokenizer2.nextToken());
         }
 
-        int cheapestPrice = price[0];
-        int answer = len[0] * price[0];
+        long cheapestPrice = price[0];
+        long answer = len[0] * price[0];
         for (int i = 1; i < N - 1; i++) {
             cheapestPrice = Math.min(cheapestPrice, price[i]);
             answer += len[i] * cheapestPrice;
